requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==2.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.5.2
django-extensions==2.1.6
django-guardian==1.5.0
django-rest-knox==3.6.0
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-guardian==0.1.1
html5lib==1.0.1
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pycparser==2.19
pydotplus==2.0.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
reportlab==3.5.21
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
webencodings==0.5.1
xhtml2pdf==0.2.3
